Fix agent host diff file paths
Legacy session-db content URIs encode the blob location in the URI path,
so the original side of a diff showed a '.../before/<file>' label and was
flagged as a rename. Canonicalize them into the current layout at the
single consuming choke point, normalizeFileEdit.

Agent host URIs also rendered as '/c:/Code/...' next to 'C:\Code\...' for
the same file, because the label formatter had no drive letter
normalization and a hardcoded posix separator. Register an OS-aware
formatter for the in-process host and normalize drive letters everywhere.
