+++ src/vs/platform/agentHost/common/agentHostFileSystemService.ts
@@ -4,13 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Disposable, IDisposable } from '../../../base/common/lifecycle.js';
+import { OS } from '../../../base/common/platform.js';
 import { IFileService } from '../../files/common/files.js';
 import { InMemoryFileSystemProvider } from '../../files/common/inMemoryFilesystemProvider.js';
 import { InstantiationType, registerSingleton } from '../../instantiation/common/extensions.js';
 import { createDecorator } from '../../instantiation/common/instantiation.js';
 import { ILabelService } from '../../label/common/label.js';
 import { AgentHostFileSystemProvider, type IRemoteFilesystemConnection } from './agentHostFileSystemProvider.js';
-import { AGENT_HOST_LABEL_FORMATTER, AGENT_HOST_SCHEME } from './agentHostUri.js';
+import { AGENT_HOST_LABEL_FORMATTER, AGENT_HOST_SCHEME, agentHostLabelFormatter, LOCAL_AGENT_HOST_AUTHORITY } from './agentHostUri.js';
 
 export type { IRemoteFilesystemConnection } from './agentHostFileSystemProvider.js';
 
@@ -55,7 +56,11 @@ class AgentHostFileSystemService extends Disposable implements IAgentHostFileSys
 
 		this._fsProvider = this._register(new AgentHostFileSystemProvider());
 		this._register(_fileService.registerProvider(AGENT_HOST_SCHEME, this._fsProvider));
+
+		// Two formatters: the scheme-wide fallback for hosts whose operating
+		// system is unknown, and a more specific one for the in-process host.
 		this._register(labelService.registerFormatter(AGENT_HOST_LABEL_FORMATTER));
+		this._register(labelService.registerFormatter(agentHostLabelFormatter(LOCAL_AGENT_HOST_AUTHORITY, OS)));
 	}
 
 	registerAuthority(authority: string, connection: IRemoteFilesystemConnection): IDisposable {
