+++ src/vs/platform/agentHost/common/agentHostUri.ts
@@ -5,6 +5,7 @@
 
 import { decodeBase64, encodeBase64, VSBuffer } from '../../../base/common/buffer.js';
 import { Schemas } from '../../../base/common/network.js';
+import { OperatingSystem } from '../../../base/common/platform.js';
 import { URI } from '../../../base/common/uri.js';
 import type { ResourceLabelFormatter } from '../../label/common/label.js';
 
@@ -151,13 +152,47 @@ export function agentHostAuthority(address: string): string {
 }
 
 /**
- * Label formatter for {@link AGENT_HOST_SCHEME} URIs. The URI path is
- * already the original resource path, so the label is the path verbatim.
+ * Authority of the in-process agent host. It always runs on the same
+ * machine — and therefore the same operating system — as the client.
+ */
+export const LOCAL_AGENT_HOST_AUTHORITY = 'local';
+
+/**
+ * Fallback label formatter for {@link AGENT_HOST_SCHEME} URIs of hosts
+ * whose operating system is unknown. The URI path is already the original
+ * resource path, so the label is the path verbatim, except that a Windows
+ * drive letter is normalized (`/c:/a` renders as `C:/a`). The separator is
+ * left as `/` because a POSIX host may well be reached from a Windows
+ * client, where `\` would corrupt its paths.
  */
 export const AGENT_HOST_LABEL_FORMATTER: ResourceLabelFormatter = {
 	scheme: AGENT_HOST_SCHEME,
 	formatting: {
 		label: '${path}',
 		separator: '/',
+		normalizeDriveLetter: true,
 	},
 };
+
+/**
+ * Label formatter for {@link AGENT_HOST_SCHEME} URIs of an agent host whose
+ * operating system is known, so its paths render the way they do natively
+ * on that host (`C:\a\b` instead of `/c:/a/b`). This matters because such
+ * URIs are shown side by side with plain `file:` URIs — for example as the
+ * original side of a diff — which are always rendered natively.
+ *
+ * Registered per authority since the operating system cannot be derived
+ * from the client: a Windows client may be connected to a POSIX host.
+ */
+export function agentHostLabelFormatter(authority: string, os: OperatingSystem): ResourceLabelFormatter {
+	const isWindows = os === OperatingSystem.Windows;
+	return {
+		scheme: AGENT_HOST_SCHEME,
+		authority,
+		formatting: {
+			label: '${path}',
+			separator: isWindows ? '\\' : '/',
+			normalizeDriveLetter: isWindows,
+		},
+	};
+}
